Add a compressed word type
This is one of the most used types in the replay files. Making a separate type for it makes it easier to prove that it's correct. It also makes it easier to benchmark.

A lot of things need to be switched over to using it. I started with the actor IDs because they were the lowest impact change.